(fix) Ensure runtime and napi are emitted as needed.
Per issue #79 there seems to be a regression introduced
in 7b6dcbd which caused the `target.runtime` to no longer
be emitted when a name is used. A name is also _always_
used even if not passed, as it will grab a name from the
`package.json`. This commit also dropped including the
`napi` tag, which is still utilized downstream.

This change fixes the regression and follows this logic;

1. Emit `name` as a tag (this will always be true due to
`addName` function logic)
2. Emit `runtime` as a tag always
3. Emit `napi` as a tag if option enabled (likely always true,
as it if defaulted to true)